backend/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
aiosqlite==0.19.0
docker==7.0.0
jinja2==3.1.3
python-multipart==0.0.6
google-cloud-storage==2.14.0
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi-sso==0.9.0
mcstatus==11.1.1
python-a2s==1.3.0
python-rcon==0.5.1
